Working sleeping hunger and passive hunger (bugged)
